requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
requests==2.25.0
psycopg2==2.8.6
google-cloud-firestore==2.0.2

